Fixed sampler

+++ Audio/DSP/Filter.ipp
@@ -54,7 +54,7 @@ inline void Audio::DSP::Filter::GenerateFilterLowPass(const FIRSpec specs, float
     const std::size_t first = size / 2;
 
     for (auto i = 0ul; i < size; ++i) {
-        float idx = static_cast<float>(i - first);
+        float idx = static_cast<float>(static_cast<int>(i) - static_cast<int>(first));
         float coef = *window;
         if constexpr (ProcessWindow)
             coef = ComputeWindow<false>(specs.windowType, i, size, true);
